+++ site/src/pages/WorkspaceSettingsPage/WorkspaceSharingPage/UserOrGroupAutocomplete.tsx
@@ -3,12 +3,11 @@ import { users } from "api/queries/users";
 import type { Group, User } from "api/typesGenerated";
 import { Autocomplete } from "components/Autocomplete/Autocomplete";
 import { AvatarData } from "components/Avatar/AvatarData";
-import { isGroup } from "components/UserOrGroupAutocomplete/UserOrGroupOption";
 import { Check } from "lucide-react";
 import { type FC, useState } from "react";
 import { keepPreviousData, useQuery } from "react-query";
 import { prepareQuery } from "utils/filters";
-import { getGroupSubtitle } from "utils/groups";
+import { getGroupSubtitle, isGroup } from "utils/groups";
 
 type AutocompleteOption = User | Group;
 export type UserOrGroupAutocompleteValue = AutocompleteOption | null;
