+++ site/src/utils/groups.ts
@@ -1,4 +1,16 @@
-import type { Group } from "api/typesGenerated";
+import type { Group, ReducedUser, User } from "api/typesGenerated";
+
+type UserOrGroupAutocompleteValue = User | ReducedUser | Group | null;
+
+/**
+ * Type guard to check if the value is a Group.
+ * Groups have a "members" property that users don't have.
+ */
+export const isGroup = (
+	value: UserOrGroupAutocompleteValue,
+): value is Group => {
+	return value !== null && typeof value === "object" && "members" in value;
+};
 
 /**
  * Returns true if the provided group is the 'Everyone' group.
